+++ src/App.js
@@ -1,26 +1,42 @@
-import logo from './logo.svg';
-import './App.css';
+import { Outlet, Link } from "react-router-dom";
+
 
 function App() {
   return (
-    </div >
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+    <div>
+
+      <nav className="navbar navbar-expand-lg navbar-light bg-light">
+        <a className="navbar-brand" href="./">Navbar</a>
+        <button className="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+          <span className="navbar-toggler-icon"></span>
+        </button>
+
+        <div className="collapse navbar-collapse" id="navbarSupportedContent">
+          <ul className="navbar-nav mr-auto">
+            <li className="nav-item active">
+              <a className="nav-link" href="./">Home <span className="sr-only">(current)</span></a>
+            </li>
+
+            <li className="nav-item">
+              <a className="nav-link" href="login">Login</a>
+            </li>
+            <li className="nav-item">
+              <a className="nav-link" href="inventory">Inventory</a>
+            </li>
+            <li className="nav-item">
+              <a className="nav-link" href="Input">Input</a>
+            </li>
+
+
+          </ul>
+
+        </div>
+      </nav>
+
+      <Outlet />
+
     </div>
+
   );
 }
-
 export default App;
