+++ src/components/Input.js
@@ -1,0 +1,21 @@
+
+function Input() {
+    return (
+        <div>
+            <h1>Inventory Input Page</h1>
+
+            <label>Product Id #: </label>
+            <br />
+            <button> - </button>
+            <input placeholder="procuct id #"></input>
+            <button>+</button>
+            <br />
+            <label>Quantity</label>
+            <br />
+            <button>-</button>
+            <input placeholder="How many?"></input>
+            <button>+</button>
+        </div>
+    );
+}
+export default Input;
