+++ src/components/Inventory.js
@@ -1,0 +1,30 @@
+export default function Inventory() {
+    return (
+        <div>
+            <h1>Inventory Page</h1>
+            <table>
+                <tr>
+                    <th>Item Name</th>
+                    <th>Product #</th>
+                    <th>Quantity In Stock</th>
+                </tr>
+                <tr>
+                    <td>Galaxy</td>
+                    <td>123456</td>
+                    <td>10</td>
+                </tr>
+                <tr>
+                    <td>iPhone</td>
+                    <td>654321</td>
+                    <td>5</td>
+                </tr>
+                <tr>
+                    <td>BlueBerry</td>
+                    <td>987654</td>
+                    <td>100</td>
+                </tr>
+
+            </table>
+        </div >
+    );
+}
