+++ src/components/Login.js
@@ -1,11 +1,10 @@
-import React from "react";
-import "./Login.css";
-
 function Login() {
     return (
         <div>
+            <h1>Login Page</h1>
             <label>Username:  </label>
             <input placeholder="username"></input>
+            <br />
             <label>Password:  </label>
             <input placeholder="password"></input>
         </div>
