--- src/screens/Input.js
@@ -1,14 +1,0 @@
-import React from "react";
-import "./Input.css";
-
-function Input() {
-    return (
-        <div>
-            <h1>Inventory Input Page</h1>
-            <label>Product Id #: </label>
-            <input></input>
-
-        </div>
-    );
-}
-export default Input;
