--- src/screens/Inventory.js
@@ -1,15 +1,0 @@
-import React from 'react'
-
-export default function Inventory() {
-    return (
-        <div>
-            <h1>Welcome to the Inventory Page! :D</h1>
-            <h2>Info this page should display</h2>
-
-        </div>
-    );
-}
-
-/*export default Inventory;
-
-*/
